chore: move util functions in Bolt Core to standalone files; update core Preact and HyperHTML library dependencies being exported
